+++ CHANGELOG.md
@@ -36,6 +36,8 @@
   * `url`:
     * Added `is_shortened_url` which checks if an URL is a short link by comparing it to a list of popular services. (Currently 24 domains recognized)
     * Renamed `normalize_query_part` to `_normalize_query_part` to mark it as an internal helper function.
+* Other:
+  * Changed all severity levels of logging messages from error / warning to debug. The package must not spam the logs of the application using it.
 
 
 ## Version 1.0.0 (2023-10-10)